requirements.txt
dataclasses-json
rawpy
tqdm
exifread
Pillow